requirements.txt
vtk
Pillow
scipy
pydicom
scikit-image
natsort
tensorflow
tensorflow_addons
tflite-runtime
scikit-learn
testresources
open3d
vedo
cylinder_fitting
gspread
oauth2client
psutil
dash
dash-bootstrap-components
plotly
tqdm
numpy
PyYAML
nibabel
batchgenerators
SimpleITK
black
flake8
isort